+++ backend/memories/signals.go
@@ -7,7 +7,7 @@ import (
 
 var MemoryCreatedSignal = signals.New[*Memory]()
 var MemoryUpdatedSignal = signals.New[*Memory]()
-var MemoryDeletedSignal = signals.New[int64]()
+var MemoryDeletedSignal = signals.New[int]()
 
 var MemoryPreferencesUpdatedSignal = signals.New[*MemoryPreferences]()
