+++ backend/worlds/signals.go
@@ -7,7 +7,7 @@ import (
 
 var WorldCreatedSignal = signals.New[*World]()
 var WorldUpdatedSignal = signals.New[*World]()
-var WorldDeletedSignal = signals.New[int64]()
+var WorldDeletedSignal = signals.New[int]()
 
 var ChatPreferencesUpdatedSignal = signals.New[*ChatPreferences]()
